atualizar , controller, service

+++ src/main/java/io/github/spring/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.spring.libraryapi.service;
 
+import io.github.spring.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.spring.libraryapi.model.Autor;
 import io.github.spring.libraryapi.repository.AutorRepository;
+import io.github.spring.libraryapi.repository.LivroRepository;
+import io.github.spring.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,21 +16,28 @@ import java.util.stream.Collectors;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-
-    public AutorService(AutorRepository repository) {
+    public AutorService(AutorRepository repository,
+                        AutorValidator validator,
+                        LivroRepository livroRepository) {
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
+        validator.validar(autor);
         return repository.save(autor);
     }
 
     public void atualizar(Autor autor) {
-        if(autor.getId() == null) {
+        if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar, é necessario que o autor já esteja salvo na base");
         }
-         repository.save(autor);
+        validator.validar(autor);
+        repository.save(autor);
     }
 
     public Optional<Autor> obterPorId(UUID id) {
@@ -35,6 +45,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if(possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um Autor que possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -58,5 +72,9 @@ public class AutorService {
                     .collect(Collectors.toList());
         }
         return autores;
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
